Feat: First working chat model
This is the first history aaware chat model that works pretty
well for the intended purposes. It contains two chains
The first chain rewrites the query that the user provides
using the history of dialogues. This query is then
given to the retriever to retrieve relevant parts of the store
before answering the question using large language models.