Feat: Add functionality to read pdf documents
Added functionality to ingest pdf documents and split them
into different sections. This just needed us to add
a pdf document reader. I added Unstructured as the pdf document
loader. I also tried pypdf. But this one seems to read
it better compared to pypdf in terms of what is read.
From this you can create another chain that answers
questions about the pdf. But the next step would be
to create these chains to tools. A llm should be able
to route the different queries to different tools
and answer them. For example, my responsibilities
at the company are not present in the website.
My CV dos not have information about my blogposts.