Feat: Add a feature complete app
This app can accept the zip folder for notion and take a CV
Then answer questions. Many improvements like streaming
are added to the app. This can be considered feature complete
But there are many improvements that can be made. I will make the
UX improvements like adding timer icon while loading the chain
or waiting for the agent. The agent takes too much time
to execute actually. This might be a draw back of the system
design. A better way to do this might be multivectorstore
Just get the information from other vectorstores.